Harmony Patches
* Adds harmony patches for ThingSetMaker so if they spawn hide it will be a single hide with proper leather amount inside.

+++ 1.5/Source/MedievalOverhaul/Medieval Overhaul/ThingClasses/DefModExtensions/FuelValueProperty.cs
@@ -5,7 +5,7 @@ namespace MedievalOverhaul
 {
     public class FuelValueProperty : DefModExtension
     {
-        public int fuelValue;
+        public float fuelValue;
 
         public static FuelValueProperty Get(Def def) => def.GetModExtension<FuelValueProperty>();
     }
